vk: Move renderer types to a separate file
- Makes my life easier managing conflicts
